requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
python-multipart==0.0.20
PyJWT==2.10.1
sqlalchemy==2.0.43
psycopg2==2.9.10
python-dotenv==1.1.1
passlib==1.7.4
bcrypt==4.3.0